Migrate HealthEndpoint to Tapi
