+++ webapi/src/main/scala/org/knora/webapi/slice/common/api/DocsGenerator.scala
@@ -33,7 +33,7 @@ import org.knora.webapi.slice.admin.api.ProjectsEndpoints
 import org.knora.webapi.slice.admin.api.StoreEndpoints
 import org.knora.webapi.slice.admin.api.UsersEndpoints
 import org.knora.webapi.slice.admin.domain.model.User
-import org.knora.webapi.slice.infrastructure.api.VersionEndpoint
+import org.knora.webapi.slice.infrastructure.api.ManagementEndpoints
 import org.knora.webapi.slice.resourceinfo.api.ResourceInfoEndpoints
 import org.knora.webapi.slice.search.api.SearchEndpoints
 
@@ -59,7 +59,7 @@ object DocsGenerator extends ZIOAppDefault {
       _               <- ZIO.logInfo("Generating OpenAPI docs")
       args            <- getArgs
       adminEndpoints  <- ZIO.serviceWith[AdminApiEndpoints](_.endpoints)
-      versionEndpoint <- ZIO.serviceWith[VersionEndpoint](_.endpoints)
+      versionEndpoint <- ZIO.serviceWith[ManagementEndpoints](_.endpoints)
       v2Endpoints     <- ZIO.serviceWith[ApiV2Endpoints](_.endpoints)
       path             = Path(args.headOption.getOrElse("/tmp"))
       filesWritten <-
@@ -83,7 +83,7 @@ object DocsGenerator extends ZIOAppDefault {
     SearchEndpoints.layer,
     StoreEndpoints.layer,
     UsersEndpoints.layer,
-    VersionEndpoint.layer,
+    ManagementEndpoints.layer,
   )
 
   private def writeToFile(endpoints: Seq[AnyEndpoint], path: Path, name: String) = {
